+++ src/app.js
@@ -1,13 +1,22 @@
 import './lib/telegram';
 import express  from 'express';
-import router   from './router';
+import router   from 'controllers/express/router';
+import middlewares from 'controllers/express/middlewares';
+import logger from 'lib/logger';
 import config   from './config';
 
-const { port, prefix } = config;
+const { port, prefix } = config.app;
 const app = express();
+
+app.use(middlewares.json);
+app.use(middlewares.urlencoded);
+app.use(middlewares.cors);
+app.use(middlewares.arrays);
 
 app.use(prefix, router);
 
 app.listen(port, () => {
-    console.log(`APP STARTING AT ${port} PORT`);
+    logger.log(`APP STARTING AT ${port} PORT`);
 });
+
+export default app;