+++ src/errors/X.js
@@ -1,10 +1,13 @@
 import { inspect } from 'util';
+import errors, { unknown } from 'constants/errors';
 
 export default class X extends Error {
-    constructor() {
+    constructor(error) {
         super();
         this.name = this.constructor.name;
         Error.captureStackTrace(this, this.constructor);
+        this.stack = error.stack;
+        this.message = error.message;
     }
 
     static stringify(data) {
@@ -15,7 +18,12 @@ export default class X extends Error {
         });
     }
 
-    get stack() {
-        return this.parent.stack;
+    render() {
+        const message = unknown;
+
+        return {
+            code : errors[message],
+            message
+        };
     }
 }